refactor(data): move buildSearchBarLayout to data/layout/ (refs #298)
