+++ src/data/layout/buildSearchBarLayout.ts
